fire-scanner/requirements.txt
faster-whisper
webrtcvad
requests
python-dotenv
